Set up character tagging
Made the character tagging work as intended.

You have 5 characters, after clicking on each on a menu pops up for you
to choose which character you are guessing that is.
The character you have guessed, the coordinates you have clicked, as well
as the correct coordinates of the character you are guessing are stored
in state.

If the coordinates you have clicked and the coordinates of the guessed
character are matching, the character is marked as located in state.

After all 5 correct guesses, a state change occurs which notifies that
all characters are found.